+++ src/js/util/player.js
@@ -81,7 +81,10 @@ export class Player {
             this.enableApi().then(() => post(this.el, {func: 'playVideo', method: 'play'}));
         } else if (this.isHTML5()) {
             try {
-                this.el.play();
+                const promise = this.el.play();
+                if (promise) {
+                    promise.catch(noop);
+                }
             } catch (e) {}
         }
     }