+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Fix setting height of drag in Sortable component
 - _isInView_ checks for element visibility now
 - Fix check for backgroundSize: cover resetting backgroundSize in Parallax
+- Fix "Uncaught (in promise) DOMException" in Player
 
 ## 3.0.0 beta 40 (February 8, 2018)